Wip de dispach con promise
